requirements.txt
aiogram==2.6.1
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
idna==2.9
multidict==4.7.5
psycopg2-binary==2.8.4
pyTelegramBotAPI==3.6.7
pytz==2019.3
requests==2.7.0
six==1.14.0
yadisk==1.2.14
yarl==1.4.2
